Initial commit
This is a fork of @remirror/extension-yjs 1.0.15, adjusted to build standalone and have passing tests. Note that
the build output is in ESM format, targeting ES2019. Standard remirror packages target a wide range of environments,
but for now the complexity of setting that up isn't worth it for our use case.