Use a single `UndoManager` instance for the life-time of the Yjs extension
The `yUndoPlugin` by default creates the undo manager based on the `protectedNodes` and
`trackedOrigins` as part of setting up a "state", but that leads to memory leaks as the
undo manager itself attaches itself as listener to the Y.Doc `afterTransaction` event.

By creating the undo manager on the outside we avoid that problem.

Relates-to (hopefully fixes...) https://github.com/Collaborne/backlog/issues/238